update user profile

+++ routes/users.js
@@ -102,35 +102,41 @@ router.post('/updateProfile', function (req, res) {
               res.statusCode = 500;
               res.send(errorMessage(err));
             }else{
-              var sql2 = "delete from adv.interests where userid = ?";
-              mysql.execQuery(sql2,[user_id],function(err, results){
-                if(err){
-                  res.status(500);
-                  res.send({"msg": "error updating users", "status":"fail", data: results});
-                  res.end();
-                }else{
-                  var length = interests.length;
-                  var temp = 0;
-                  var sql3 = "insert into interests (userid,category_id,category_name) values(?,?,?);"
-                  for(var i = 0 ; i < length ; i ++) {
-                    var params = [user_id, interests[i].category_id, interests[i].category_name];
-                    mysql.execQuery(sql3,params,function(err, results){
-                      if(err){
-                        res.status(500);
-                        res.send({"msg": "error updating interests", "status":"fail", data: results});
-                        res.end();
-                      }else{
-                        temp++;
-                        if(temp === length){
-                          res.status(200);
-                          res.send({"msg": "updated user","status":"success"});
+              if(interests.length) {
+                var sql2 = "delete from adv.interests where userid = ?";
+                mysql.execQuery(sql2,[user_id],function(err, results){
+                  if(err){
+                    res.status(500);
+                    res.send({"msg": "error updating users", "status":"fail", data: results});
+                    res.end();
+                  }else{
+                    var length = interests.length;
+                    var temp = 0;
+                    var sql3 = "insert into interests (userid,category_id,category_name) values(?,?,?);"
+                    for(var i = 0 ; i < length ; i ++) {
+                      var params = [user_id, interests[i].category_id, interests[i].category_name];
+                      mysql.execQuery(sql3,params,function(err, results){
+                        if(err){
+                          res.status(500);
+                          res.send({"msg": "error updating interests", "status":"fail", data: results});
                           res.end();
+                        }else{
+                          temp++;
+                          if(temp === length){
+                            res.status(200);
+                            res.send({"msg": "updated user","status":"success"});
+                            res.end();
+                          }
                         }
-                      }
-                    });
+                      });
+                    }
                   }
-                }
-              });
+                });
+              }else{
+                res.status(200);
+                res.send({"msg": "updated user","status":"success"});
+                res.end();
+              }
             }
           });
         }else{
